refactor: complete stage 2 - extract interfaces (rpc, web, bot)
