+++ src/backfill/taiwan_stock_index.rs
@@ -1,9 +1,9 @@
 use anyhow::Result;
 use chrono::{Local, NaiveDate, TimeZone};
 
-use crate::bot::telegram::Telegram;
+use crate::interfaces::bot::telegram::Telegram;
 use crate::core::util::map::Keyable;
-use crate::{bot, cache::SHARE, crawler::twse, database::table, core::logging};
+use crate::{interfaces::bot, cache::SHARE, crawler::twse, database::table, core::logging};
 
 /// 解析單筆指數字串陣列。若格式錯誤或解析失敗，會記錄 error log 並回傳 `None`。
 fn parse_index_item(item: &[String]) -> Option<table::index::Index> {
