+++ src/interfaces/mod.rs
@@ -1,0 +1,3 @@
+pub mod bot;
+pub mod rpc;
+pub mod web;
