+++ src/interfaces/rpc/stock.rs
